move table schema information to own module
